+++ rts/Rendering/Textures/Bitmap.h
@@ -12,7 +12,7 @@
 #endif // !HEADLESS
 #include "System/float3.h"
 #include "System/Color.h"
-#include "Rendering/Textures/Texture.hpp"
+#include "Rendering/Textures/TextureCreationParams.hpp"
 
 
 struct SDL_Surface;
